fix board validation and give option to force opponent move if they disconnect + simplify checkers screen state

+++ app/src/main/java/io/silv/checkers/navigation/GameNavigation.kt
@@ -3,24 +3,40 @@ package io.silv.checkers.navigation
 import android.os.Parcelable
 import android.util.Log
 import androidx.activity.compose.BackHandler
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.text.selection.SelectionContainer
+import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.window.Popup
 import com.bumble.appyx.core.composable.Children
 import com.bumble.appyx.core.modality.BuildContext
 import com.bumble.appyx.core.node.Node
 import com.bumble.appyx.core.node.ParentNode
 import com.bumble.appyx.navmodel.backstack.BackStack
+import com.bumble.appyx.navmodel.backstack.operation.pop
 import com.bumble.appyx.navmodel.backstack.operation.push
+import com.bumble.appyx.navmodel.backstack.operation.singleTop
 import com.bumble.appyx.navmodel.backstack.transitionhandler.rememberBackstackSlider
+import io.silv.checkers.Cord
+import io.silv.checkers.Piece
 import io.silv.checkers.Room
 import io.silv.checkers.screens.CheckersScreen
+import io.silv.checkers.ui.ConfirmLeavePopup
 import io.silv.checkers.viewmodels.CheckerUiState
 import io.silv.checkers.viewmodels.CheckersViewModel
 import kotlinx.parcelize.Parcelize
@@ -33,7 +49,7 @@ sealed class GameNavTarget: Parcelable {
     data class Connecting(val roomId: String): GameNavTarget()
 
     @Parcelize
-    data class Queue(val roomId: String, val room: Room): GameNavTarget()
+    data class Queue(val roomId: String): GameNavTarget()
 
     @Parcelize
     data class Game(val roomId: String): GameNavTarget()
@@ -63,23 +79,15 @@ class CheckersGame(
 
         val viewModel: CheckersViewModel = koinViewModel() { parametersOf(roomId) }
 
-        val state = viewModel.uiState.collectAsState().value
+        val state by viewModel.uiState.collectAsState()
 
-        LaunchedEffect(key1 = state) {
-            when (state) {
-                is CheckerUiState.Playing -> {
-                    Log.d("FB", "called playing joinedRoom")
-                    backStack.push(
-                        GameNavTarget.Game(roomId)
-                    )
-                }
-                is CheckerUiState.Queue -> {
-                    Log.d("FB", "called playing joinedRoom")
-                    backStack.push(
-                        GameNavTarget.Queue(roomId, state.room)
-                    )
-                }
-                else -> Unit
+        LaunchedEffect(state.room.usersToColorChoice) {
+            if (state.room.usersToColorChoice.size >= 2) {
+                backStack.pop()
+                backStack.push(GameNavTarget.Game(roomId))
+            } else {
+                backStack.pop()
+                backStack.push(GameNavTarget.Queue(roomId))
             }
         }
 
@@ -102,14 +110,10 @@ class Game(
 
         val viewModel: CheckersViewModel = koinViewModel()  { parametersOf(roomId) }
 
-        when (val state = viewModel.uiState.collectAsState().value) {
-            is CheckerUiState.Playing -> {
-                CheckersScreen(state)
-            }
-            else -> {
-                // TODO("Navigate back on error")
-            }
+        val state by viewModel.uiState.collectAsState()
 
+        CheckersScreen(state) { from: Cord, to: Cord, piece: Piece ->
+            viewModel.onDropAction(state.room, state.rawBoard, state.board, from, to, piece)
         }
     }
 }
@@ -122,27 +126,37 @@ class Queue(
 
     @Composable
     override fun View(modifier: Modifier) {
+
         val viewModel: CheckersViewModel = koinViewModel()
+        val state by viewModel.uiState.collectAsState()
 
-
-        BackHandler {
-            viewModel.deleteRoom(roomId)
-            navigateBack()
+        var leaveConfirmationVisible by remember {
+            mutableStateOf(false)
         }
 
-        when (val state = viewModel.uiState.collectAsState().value) {
-            is CheckerUiState.Queue -> {
-                Box(
-                    modifier = Modifier
-                        .fillMaxSize(),
-                    contentAlignment = Alignment.Center
-                ) {
-                    SelectionContainer() {
-                        Text(text = "waiting for opponent ${state.room}")
-                    }
-                }
+        ConfirmLeavePopup(
+            show = leaveConfirmationVisible,
+            onConfirm = {
+                viewModel.deleteRoom(roomId)
+                navigateBack()
+            },
+            onDeny = {
+                leaveConfirmationVisible = false
             }
-            else -> Unit
+        )
+
+        BackHandler {
+           leaveConfirmationVisible = true
+        }
+
+        Box(
+            modifier = Modifier
+                .fillMaxSize(),
+            contentAlignment = Alignment.Center
+        ) {
+            SelectionContainer() {
+                Text(text = "waiting for opponent ${state.room}")
+            }
         }
     }
 }
@@ -162,7 +176,7 @@ class Connecting(
                 .fillMaxSize(),
             contentAlignment = Alignment.Center
         ) {
-            SelectionContainer() {
+            SelectionContainer {
                 Text(text = "connecting $roomId")
             }
         }
